When sudo is missing, fall through to sysadmin instructions
If `sudo` is missing, print the sysadmin instructions, not the
instructions requiring presence of `sudo`.

Fixes https://github.com/clearlinux/distribution/issues/457

+++ 10-command-not-found.sh
@@ -2,7 +2,7 @@ command_not_found_handle () {
     echo "$1: command not found"
     if [ "`id -u`" == "0" ]; then
         MSG='$1 == cmd { print "To install " $1 " use:  swupd bundle-add " $2; exit }'
-    elif groups | grep -q -w -e wheel -e wheelnopw ; then
+    elif [ "$1" != "sudo" ] && groups | grep -q -w -e wheel -e wheelnopw ; then
         MSG='$1 == cmd { print "To install " $1 " use:  sudo swupd bundle-add " $2; exit }'
     else
         MSG='$1 == cmd { print "To install " $1 " your system administrator needs to do:  swupd bundle-add " $2; exit }'
